Renamed `FixturesLoader` into the `FixturesProvider` and moved all functions from the `SwaggerFactory`.
