+++ Tests/Fixtures/FixturesProvider.php
@@ -11,15 +11,30 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Linkin\Bundle\SwaggerResolverBundle\Tests;
+namespace Linkin\Bundle\SwaggerResolverBundle\Tests\Fixtures;
 
 use EXSyst\Component\Swagger\Schema;
+use EXSyst\Component\Swagger\Swagger;
 
 /**
  * @author Viktor Linkin <adrenalinkin@gmail.com>
  */
-class SwaggerFactory
+class FixturesProvider
 {
+    /**
+     * @var Swagger
+     */
+    private static $cachedSwagger;
+
+    public static function loadFromJson(): Swagger
+    {
+        if (self::$cachedSwagger === null) {
+            self::$cachedSwagger = Swagger::fromFile(__DIR__ . '/Json/customer.json');
+        }
+
+        return self::$cachedSwagger;
+    }
+
     /**
      * @param array $properties
      * @param array $required
