+++ Tests/Normalizer/IntegerNormalizerTest.php
@@ -17,7 +17,7 @@ use EXSyst\Component\Swagger\Schema;
 use Linkin\Bundle\SwaggerResolverBundle\Exception\NormalizationFailedException;
 use Linkin\Bundle\SwaggerResolverBundle\Normalizer\IntegerNormalizer;
 use Linkin\Bundle\SwaggerResolverBundle\Resolver\SwaggerResolver;
-use Linkin\Bundle\SwaggerResolverBundle\Tests\SwaggerFactory;
+use Linkin\Bundle\SwaggerResolverBundle\Tests\Fixtures\FixturesProvider;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -163,7 +163,7 @@ class IntegerNormalizerTest extends TestCase
 
     private function createSchemaDefinition(string $fieldName, string $type = self::TYPE_INTEGER): Schema
     {
-        return SwaggerFactory::createSchemaDefinition([
+        return FixturesProvider::createSchemaDefinition([
             $fieldName => [
                 'type' => $type
             ]
